Code refactor & clean up
1. Moved `getMediaInfo` utility function to `asset-services`
2. Removed unused fuctions and code clean up
