+++ app/screens/image-gallery-viewer/gallery-image.tsx
@@ -26,13 +26,13 @@ import Animated, {
   withTiming,
 } from "react-native-reanimated"
 import { SharedElement } from "react-navigation-shared-element"
+import { ScreenWidth } from "@rneui/base"
 import { Video } from "expo-av"
 
 import { Text } from "../../components"
 import { palette } from "../../theme"
 import { Asset } from "../../types"
-import { ScreenHeight, ScreenWidth } from "@rneui/base"
-import type { Video as VideoType } from "expo-av"
+import { AssetService } from "../../services"
 
 type GalleryImageProps = {
   asset: Asset
@@ -46,24 +46,6 @@ type GalleryImageProps = {
 const MAX_SCALE = 6
 const SWIPE_UP_THRESHOLD = 10
 const SWIPE_TO_CLOSE_THRESHOLD = 100
-
-import * as MediaLibrary from "expo-media-library"
-import type { Asset as Media } from "expo-media-library"
-
-const getUserAssets = async (): Promise<Media[] | null> => {
-  const { assets } = await MediaLibrary.getAssetsAsync({
-    mediaType: ["photo", "video", "audio", "unknown"],
-  })
-  if (!assets) {
-    return null
-  }
-  return assets
-}
-
-const getMediaInfo = async (asset: Asset): Promise<MediaLibrary.AssetInfo> => {
-  const info = await MediaLibrary.getAssetInfoAsync(asset)
-  return info
-}
 
 export const GalleryImage: React.FC<GalleryImageProps> = ({
   asset,
@@ -353,9 +335,9 @@ export const GalleryImage: React.FC<GalleryImageProps> = ({
   const [, setPlaybackStatus] = React.useState({})
 
   const getAssetLocalInfo = React.useCallback(async () => {
-    const localInfo = await getMediaInfo(asset)
+    const localInfo = await AssetService.getMediaInfo(asset)
     setLocalUri(localInfo.localUri)
-  }, [getMediaInfo, asset])
+  }, [AssetService.getMediaInfo, asset])
 
   React.useLayoutEffect(() => {
     getAssetLocalInfo()
