+++ app/services/asset-service.ts
@@ -2,6 +2,8 @@ import moment from "moment"
 import * as MediaLibrary from "expo-media-library"
 import { manipulateAsync, SaveFormat, ImageResult } from "expo-image-manipulator"
 
+import type { Asset as AssetType } from "expo-media-library"
+
 import { RecyclerAssetListSection, ViewType, GroupHeader, Library, AssetStory } from "../types"
 export const generateThumbnail = async (assets: MediaLibrary.Asset[]) => {
   const result: ImageResult[] = []
@@ -59,7 +61,7 @@ const getAssetStoryCategory = (modificationTime: number) => {
   return null
 }
 
-export const categorizeAssets = (assets: MediaLibrary.Asset[], storyHighlight=false) => {
+export const categorizeAssets = (assets: MediaLibrary.Asset[], storyHighlight = false) => {
   const sections: RecyclerAssetListSection[] = []
   let lastMonth = moment().format("MMMM YYYY")
   let lastDay = null
@@ -70,7 +72,10 @@ export const categorizeAssets = (assets: MediaLibrary.Asset[], storyHighlight=fa
   for (const asset of assets) {
     // Make story objects
     // Filter assets to get camera folder
-    if (__DEV__ || (asset?.modificationTime && asset.uri?.toLowerCase().includes("/dcim/camera/"))) {
+    if (
+      __DEV__ ||
+      (asset?.modificationTime && asset.uri?.toLowerCase().includes("/dcim/camera/"))
+    ) {
       const categoryName = getAssetStoryCategory(asset.modificationTime)
       if (categoryName && !storiesObj[categoryName]) storiesObj[categoryName] = []
       if (categoryName) storiesObj[categoryName].push(asset)
@@ -128,7 +133,7 @@ export const categorizeAssets = (assets: MediaLibrary.Asset[], storyHighlight=fa
     }
   }
 
-  if(storyHighlight){
+  if (storyHighlight) {
     // Create story section
     const storySection: RecyclerAssetListSection = {
       id: "story_highlight",
@@ -148,7 +153,6 @@ export const categorizeAssets = (assets: MediaLibrary.Asset[], storyHighlight=fa
   }
 
   return [...sections]
-  
 }
 export const getLibraries = (assets: MediaLibrary.Asset[]): Library[] => {
   const librariesObj: Record<string, MediaLibrary.Asset[]> = {}
@@ -181,16 +185,16 @@ export const getAssets = async (
     const medias = await MediaLibrary.getAssetsAsync(
       afterAssetId
         ? {
-          first: pageSize,
-          after: afterAssetId,
-          sortBy: sortBy,
-          mediaType: ["photo", "video"],
-        }
+            first: pageSize,
+            after: afterAssetId,
+            sortBy: sortBy,
+            mediaType: ["photo", "video"],
+          }
         : {
-          first: pageSize,
-          sortBy: sortBy,
-          mediaType: ["photo", "video"],
-        },
+            first: pageSize,
+            sortBy: sortBy,
+            mediaType: ["photo", "video"],
+          },
     )
     return medias
   } catch (error) {
@@ -206,4 +210,9 @@ export const deleteAssets = async (assetIds: string[]): Promise<boolean> => {
     console.error("error", error)
     throw error
   }
+}
+
+export const getMediaInfo = async (asset: AssetType): Promise<MediaLibrary.AssetInfo> => {
+  const info = await MediaLibrary.getAssetInfoAsync(asset)
+  return info
 }
